changed conv_theta and visualize
conv_theta changed to just use unwrap, this seems to work better. I dont remember what was the exact issue with it before, but if it emerges again we can just fix it then. Also now the derivative is double filtered. So theta is filtered, then differentiated, then filtered again with same filter. Seems to give nice results.